read bayot

+++ src/main/java/com/wildteach/tutoringsystem/repository/wildteachRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.wildteach.tutoringsystem.entity.wildteachEntity;
 
 @Repository
-public interface wildteachRepository extends JpaRepository<wildteachEntity, Integer>{
+public interface wildteachRepository extends JpaRepository<wildteachEntity, Long>{
 
 }
